% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_mass.R
\name{calc_shrub_mass}
\alias{calc_shrub_mass}
\title{Calculate the mass of a shrub}
\usage{
calc_shrub_mass(len, width, height)
}
\arguments{
\item{len}{Length of the object}

\item{width}{Width of the object}

\item{height}{Height of the object}
}
\value{
volume The density of the object (numeric)
}
\description{
Use the length, width, height of an object to calculate its volume
Returns density as a floating point number.
}
